requirements.txt
torch
transformers
optimum
onnx
openai-whisper
langchain
pydub
openai
ffmpeg-python
onnxruntime-directml>=1.15.0
pillow
azure.keyvault.secrets
azure.identity
